Add a "global" turbo stream for site-wide event notifications
To support this, had to reorganize the application layout a bit, else
the turbo_stream tag was adding an additional gap (at the top of the
page) due to the flex-col styling that was on the body tag. It was
probably never really a great idea to style the body tag itself like
this... it was probably inevitable.