Change War Room connection status light -> activity indicator
The "connection status light" was more of a debugging concept. I was
hoping it would make it obvious when websocket connections were
flickering, but... it didn't really work out that way.

So, this commit refactors the green/gray indicator light into a War Room
activity indicator light.

- Green = A "Current {Game}" exists (and is {Game#on?} == true)
- Gray = No "Current {Game}" exists

To support this:
- Add Game::Current.exists?
- Add Game::Current::BroadcastWarRoomActivityIndicatorUpdateJob, which
  can check Game::Current.exists? when it runs and update the light
  one way or another.
  - Performing this in a job gets the query out of the main process and
    also gives a slightly delay / takes responsibility for morphing the
    DOM out of the service objects' hands.