Don't disconnect from WarRoom channel on Game end
Instead of disconnecting active participants from the WarRoom channel on
Game-end, we can simply block the request to refresh the page when a new
Game starts. This is less disruptive for the websocket connection, etc.

The net effect is the same. Active participants will stay on the
Just-ended Game view when another player starts a new Game. This gives
them time to sign their name, review metrics, etc. And they'll get the
Fleet Mustering notification, which is broadcast to the Just-Ended Game
footer stream.