Show "Alliance Best" awards for Past Game "Bests"
This is very similar to showing "Personal Bests" for active
participants... but applies to all Games of the given type, site-wide.

This is what I wanted, but now I'm very much feeling like we can't just
do this either-or. To do it right, we need to show both:
- Personal Bests, and
- Alliance Bests
... at the same time and in all views.

So as soon as I commit this, I'm starting on how to merge the two...