Reorg Game::Bests around a specific Game instance
... and make the queries more efficient -- the same as we just did for
User::Bests.

Also, we no longer need Game::Bests to be generic, across all Games
since we have the Game.for_bests_of_type* scopes now (as of a couple
commits ago).

It's also nice to have the parallelism with User::Bests, which has been
instance-based. The only difference is User::Bests' query methods do
need to take in a Game as context, where Game::Bests' do not. For
Game::Bests the "context" is basically "The Alliance", or "All Users".