Show Alliance/Personal Best on Just-Ended/Past Games
This commit concedes the processor/database/storage cycles needed to
just always show (on Just-Ended and Past Games) any "Bests" that may
have occurred for that Game. This wrecks our simple caching mechanism
for Past Games, but oh well. This is a much nicer UX!

Note: We're either showing an Alliance Best *or* a Personal Best
indicator in this commit. The idea being that if it's an Alliance Best
then it's also a Personal Best. However, this makes it harder to tell
when a Game was also a Personal Best... because this is only the case
when a User actively participated in the Game. So, we'll show both
indicators (where applicable) in the following commit.

This commit is otherwise a fairly big reorganization of code that undoes
some of what I've just done in the last couple of days... bringing back
responsibility for displaying metrics back into the purview of
"Past Games" (vs splitting them out between Just-Ended and Past Games).