Show Alliance and/or Personal Best on Just-Ended/Past Games
We will now show the Alliance Best *and* Personal Best indicators on
applicable Games.

This commit adjust the approach taken for alignment of these medals
(indicators) in the UI table. Instead of extending outside of the table
via negative margins, we change to putting the medals inside of a first
column on the table. We take care to ensure that if no medals are
present then the first column should take up no space. To achieve this:
- We can't have any whitespace inside of the `<td>` element, ever. So
  we wrap the HTML weirdly to make the render be the only thing that
  adds whitespace, if at all.
- We can't add padding to the `<td>` element if it's empty (so we use
  `not-empty:pr-2`
If anything adds any whitespace or non-whitespace, then the first column
of the table "hydrates" and takes up space. Huzzah!

Also, because we're no checking if a given User was an active
participant in the given Game multiple times (once for each
"Bestable Type": Score, 3BV/s, Efficiency), we add memoization to
User#active_participant_in?.