Add Most Active Minesweepers (Users) to Metrics page
We break the user rankings out by Game Type again, just like for
Best Engagements. The count is based on all active Game participation,
even if the Game wasn't a win.

We also cache the User rankings based on a key that's based on
Game.count. So we'll reload the cache every new game.

Also, increase the number of items shown in ranks from 3 to 5 (for
Best Engagements as well). Seems better to have a bit more info and
doesn't seem right to have a different size ranking list between
the two types: Best Engagements and Most Active Minesweepers. Could
change these later, though, as desired at the time.