Highlight the origin cell on HighlightNeighbors events
This just makes it more clear, especially for other participants, which
Cell is being clicked on for a Highlight event.

To support this, we switch from using "highlight[able]_neighbors" to
"highlight[able]_neighborhood"--which includes the origin cell as well.
Then, within these methods, we just mark the origin cell as the
"highlight_origin" instead of as "highlighted".

Like "highlighted", "highlight_origin" is not saved to the database. It
is just a virtual attribute.