Fix tables for Most Active Minesweepers on Metrics Page
Previously, we weren't recovering gracefully when rendering un-named
Users in the Most Active players table. Instead, we were showing the
URL, which could take up quite a bit of room, which would then cause
the table to overflow the column it was in.

So we fix both issues.
- Substitute in the User's `MMS` identifier if they don't have a
  username.
- Don't allow the table to overflow its column/container. And truncate
  the username if it's extra long.

Also, get rid of the `border-separate` and `border-spacing-*`
classes/styles as they cause weirdness. The `-mx-6` wasn't doing what
it should have... it was only nullifying the left margin. So the result
was a table that didn't take up its full col width. Instead, we switch
to doing our own cell padding, manually. Kind of a drag, but tailwind
makes it easy enough with some of the more advanced child selectors.

Also:
- Change "Count" to just "#" so the 3rd column takes up as little
  space as possible.
- Only show a browser tooltip for Users that have a username (to also
  reveal their MMS ID--via the usual User#display_name format).