Remove narrower text tracking on games-to-mines-ratio text
I think I had originally added this to give the board space for our
smallest-possible configuration at the time: Custom w/ a 4x4 grid. But
now the smallest possible is 6x6 and everything fits without reducing
the text-tracking.

Having the text-tracking reduced makes it harder to read the numbers as
they appear slightly on top of each other.