Refactor/simplify User Bests -> User Bests + Metrics
This started out as an effor to just dim the "no value indicator" on
User Show Page -> Bests -> <Metric> -> <Game Type>s. However, the code
needed to do so was fairly complex and duplicative. So, instead, I'm
starting with a refactor to first make the needed change easy. Then,
the next commit will make the easy change.

No functionality has changed here, we've just DRYed things up and split
things apart.

Also: Add an inflection for `BBBVPS` so our class names don't have to
pretend `bbbvps` is a word instead of an initialism.