Dim the text on "no value indicators"
Now we make the easy change in User Show pages, to dim the "no value
indicators", now that we've refactored User -> Bests -> Metrics in the
previous commit.

This is done to match the style from the Metrics page, which I noticed
wasn't carried over for User Bests (because it was too complex to do
before, for what it was).

Also dim the indicator in the Home Roster when no non-Guest Users are
present.