Show "away" users as dimmed in the Fleet Roster
This effort resulted in finding out that I was *still* thinking about
the Fleet Roster the wrong way the last time I revamped it. I was still
trying to get User's out of the database and use the database to order
the FleetTracker.tokens, etc. When, in fact, the database need not be
involved at all. And this realization leads to greater flexibility. So
the Fleet Tracker is actually just/only a Rails cache-based list of Game
participants now. When they go away the'll still be shown, they'll just
be dimmed to indicate that they're gone away (which is really that
they've expired).

Along the way, I've rethought, renamed, and rewritten a lot of the
FleetTracker code again. In particular, I've adjusted the public
interface to be more inline with the internal workings of the Registry,
which makes everything easier to reason about.

Also, instead of passing around or expecting User#token's for everything
related to the FleetTracker, we now pass around our Entry objects, where
possible. For example:

We can't pass an Entry object into the background job that is
responsible for expiring a User's Fleet Roster entry because, otherwise,
we can't check their current status right before broadcasting their
now-expired status. To do that, we just need a token so we can re-query
for them.
- In fact we could still pass in an Entry and get the token out of this,
  but that's not worth the pain of setting up a custom Serializer for
  encoding/decoding the Entry objects in the background job.