Move actual Rails cache writes into FleetTracker::Registry
Not sure why I suffered this for so long, as it wasn't easy to. But it
makes a ton of sense to me, as of today, to let the Registry manage its
own storage, vs the FleetTracker doing it. This way, we can simply save
and then return entries from the Registry when perform write operations.

Everything remains functionally the same (didn't even have to change the
tests).