Use <ViewModel>#to_param with turbo_stream_from
TIL the `turbo_stream_from` helper's first argument accepts any object
that responds to `to_param`. This feels like the abstraction I've been
looking for as it allows me to punt on figuring out what the ideal name
is for these things. Plus it reads better in the view code, IMO.

Reference:
https://binarysolo.blog/is-your-action-cable-connection-secure-when-using-turbo/