Move dev overlay -> Dev Portal; Make caching indicator a btn
Also:
- Just show the dev overlay as the current break point. Reveal the rest
  of the overlay on mouse-over. Collapse again on mouse-out.
- On reveal, also increase text size, remove reduced opacity, ...

This is the start of the Dev Portal. I have long wanted to move the
Pattern creation tool into a Dev Portal, so that may come soon as well.
But for now we've just created it as a viable namespace for the
_dev_overlay partial and for the ToggleDevCachingController.

Note: The redirect after toggling Dev Caching isn't always going to
work the best because part of the toggle is a server restart. If the
page looks weird after, can just refresh again. I tried adding a `sleep`
statement, but this just causes the restart to wait until after the
`sleep` to trigger as well.