Extract helpers for making User's name Turbo updateable
Previously, we were handling both the broadcast code and the view code
put in place as targets for that broadcast piecemeal.

Now:
1. Extract ApplicationHelper helpers for defining the standard HTML tags
   needed for update of User#display_name and User#username on
   broadcast.
2. Create a dedicated service object for sending that broadcast to the
   appropriate targets.

Then, we apply the new approach to the rest of the site. This commit
updates all (so far as I know) displayed usernames except for:
- Browser Page Titles
- Metrics -> Minesweepers -> Most Active tables
  - The behavior here is unique, so needs a custom solution. TBA
... to adopt this new standard. Including the "Signed, <username>"
in-place edit link on just-ended Games.
- This allows multiple browsers to be open to the same view for the
  same User and all "Signed, <username>" links will still be updated to
  reflect the latest username. Seamless!