Make User#username inline-editable in Account Show page
The option to edit one's username still only appears after a User
becomes a signer, but... for convenience, Users can now edit their
username from their account page. This is otherwise supposed to be the
same experience as editing username (a.k.a. signing) so from the
just-ended Game view.

Also:
- Some considered Rubocop/reek ignores that will be added to gemwork
  soon.
- Streamlined the approach for inline edit. Detailed in
  app/javascript/controllers/inline_edit_controller.js. (Will update
  other implementations soon, as well.)