Decouple Cell (De)highlight Controllers from ActionBehaviors
We've bundled too much behavior into the various Cell Actions via the
convenience of just including
Games::Current::Board::Cells::ActionBehaviors and then everything just
works™. But I'm feeling the pull to decouple and simplify.

We start with the Cell Highlight/Dehighlight Neighbors actions. What's
special about these is:

1. Since these actions don't write to the database, we don't *need* a
   Current User record to tie them to.
- So, since we don't need a User, let's not require one and/or attempt
  to create one for these actions. This trims the action down.

2. Also since these actions don't write to the database, we don't really
   need to worry about the "lone wolf" issue of doing an operation on a
   Cell for a Game that is no longer in progress.
   - i.e. It's fine to sacrifice this check in favor of a little speed
     for this action.

3. Cell Highlight can be easily spammed, so we want to customize the
   rate limiter here.
3b. Whereas... Cell Dehighlight should always work, regardless.
    (We don't want to leave the highlighted Cells around just because of
    an arbitrary rate limit.)

To support all of this, we create a WarRoom::Responder object, which:
wraps up the request cycle of a controller by
1. Broadcasting, and
2. Responding with
... the given `html` (joined Turbo Stream Templates)