Fix "nav too tall" issues after: Tolerate long usernames ...
I didn't notice before, but with the changes from the commit:
"Tolerate long usernames in top-right (User) nav", we introduced an
expanded-height nav bar for participating Users vs Guests.

Apparently there are differences in the way `ul > li > *` element
heights affects the overal `ul` height vs that of `div > *` element
heights. The `div` will expand with the child elements while the `ul`
will not. I'm not sure why, exactly... just some default rendering
differences.

So, to ensure the user nav link stays the same height as the rest of the
nav links, we use a `ul` there as well, even though it's just a list of
one.

This prevents the nav bar height from expanding, unnecessarily, the
moment a new User is created. Or, vice versa when "logging out".