Create GeneratePassiveResponse object
This is the other side of the coin from the GenerateActiveResponse
object we just created in the previous commit. GeneratePassiveResponse
doesn't require a Current User (a Guest can perform the action).
Currently, this applies only to Highlight/Dehighlight of Cell
"neighborhoods".

Then, we simplify the RevealNeighborsController#create action by
abstracting the dehighlighting of cell neighborhoods into
Games::Current::Board::Cells::DehighlightNeighbors.