Create/Use DispatchAction & DispatchEffect for Cell Actions
Now that I can see the forest for the trees, it's finally clear to me
what the proper abstraction level is for handling active vs passive
Cell Actions. WHat I've distilled this down to is that we're either
dispatching an "action" or we're dispatching an "effect".

TL;DR:
DispatchAction is meant for handling active Cell Actions, which affect
the {Game} play history, while
DispatchEffect is meant for handling passive Cell Actions, which don't
affect the {Game} play history.

DispatchAction:
1. Ensures an active participant ({User}) exists,
2. Yields a {Games::Current::Board::Cells::DispatchAction::Dispatch}
   object for bundling action updates together and for defining
   {Game#status}-based hooks,
3. Builds a primary Action response based on the current {Game#status},
4. Bundles together all of the Turbo Stream Actions generated
   throughout, and
5. Hands over the bundled Actions to {WarRoom::Responder} for
   broadcast/delivery.

DispatchEffect:
1. Yields a {Games::Current::Board::Cells::DispatchEffect::Dispatch}
   object for bundling action updates together,
2. Builds a primary Action based on the effect, and
3. Hands over the bundled Actions to {WarRoom::Responder} for
   broadcast/delivery.

In this commit, we build out the above and adapt these for use within
all of the Cell Actions.

We also remove the lead-up attempts at solving the same problem:
- Games::Current::Board::Cells::GenerateActiveResponse
- Games::Current::Board::Cells::GeneratePassiveResponse
... I wanted to keep the previous attempts here, in revision history,
for reflection and growth.

We also adapt Cell::ToggleFlag to conform to the expected API for use
with Games::Current::Board::Cells::DispatchAction.