Eager load Board in Cell ActionBehaviors
This is just a tiny optimization to get both the Game and the Board in
the same query when performing Cell Actions. This just saves a few
milliseconds in practice.

I also tried eager loading the Board's Cells, but this slows things
down by about 3x so leaving it out.