Add User::Console.recover for auth link lookup
This is mainly for me during testing, but... it's handy to be able to
get the Authentication Link by ... something that is known about a User.
Any of these will work:
- An MMS ID, as either "MMS-<1234>", "1234", or 1234
- A User Token, e.g. "08cd12b8-495c-4fe8-8f60-7c8b606cec53"
- A username, e.g. "pdobb" or some substring, like "pd"

A collection of results will be returned with all 3 values identified
to help determine which authentication link should be used from there.

To support this, we add a `for_mms_id` scope to User::Console. However,
it should be noted that this does add the scope  to User as well. I'm
find with this because... while I don't expect the application code to
use this scope, it's good to have around for general use. And, being
defined within the User::Console, like it is, makes it clear that it's
primary purpose is for Console use (i.e. the application code won't be
calling it ever).

Liking this approach, we also move the existing `for_tokenish` and
`for_username` scopes down to User::Console and rename them to
`like_tokenish` and `like_username` respectively. These scopes were also
never (expected to be) used by application code.