Don't auto-timeout Fleet Mustering notification
We can get a nice effect where we only ever have one Fleet Mustering
notification even if multiple are delivered--automatically, thanks to
how Turbo's prepend (and append) actions work. By giving this
notification an `id`, Turbo will automatically "remove the duplicate
target child" instead of simply prepending the new notification (with
the same ID).

Also, since the "Fleet Mustering!" notification will hang around now, it
will be more tempting to use `esc` to dismiss it. But, doing so closes
the slide menu for the Roster as well. So, we remove the `esc` button
effect on slide menus in exchange for sticky Fleet Mustering
notifications.