Fix: Don't replace just-ended Game view on 2nd+ new Game
Previously, since we kept the turbo frame tag around, the 2nd time a
Game had just-ended, all users still sitting on the first just-ended
view of a Game would have their Game container replaced as well. This
wasn't intended and is an easy fix... We don't even need to include
the turbo frame tag in the games/just_ended/_container partial... which
then prevents it from being replaced with updated content later.

Also, with this change, we also back out and clean up the back-end code
surrounding it.