Add Interaction model + Record Interaction service
This is a first attempt at a means of keeping track of user interactions
with controls, links, resources... This is not meant to be a
"page view" tracker (we will create another model for that if needed),
but more of an individualized, bespoke page element, link, resource...
tracker.

To support this, we use page_controller.js to fire off a POST request
to a new InteractionsController#create action on an interaction of
interest.

The first interaction is: "Toggle Rules of Engagement". More to come...
Usage example:

```
<button
    data-action="page#recordInteraction"
    data-interaction-name="<Feature Name>"
    ...>
```

Dev environment only: View the current counts by visiting the dev
portal. Else, for production, use the Rails console:
```
App.introspect
```

Note: We ignore Guests and "dev" Users from interaction counts.