Add User#dev column + Show indicate dev User in UI
It's useful to be able to exclude devs (myself) from interactions
(upcoming) / stats, as needed. So that is the primary motivation for
this feature. However, I also think it's kind of a cool thing for people
to be able to see in the UI, so I've added an emoji/flag for this in
the User Profile page.

To save on database resources, we only query for current "dev" User
tokens once per session. Restart the user to reload after making a
change. Adding myself as a dev looks like, e.g.:

```
User.like_username("pdobb").take!.update(dev: true)
```