Add Visit model + Record Visit service
This is a first attempt at a means of keeping track of page visits--at
the level of unique paths, and as recorded by Controller Actions. To
use, just add this to a controller of interest:

```
 after_action RecordVisit

 # OR:
 after_action RecordVisit, only: %i[index create]
```

Dev environment only: View the current counts by visiting the dev
portal. Else, for production, use the Rails console:
```
App.introspect
```

Note: We ignore Guests and "dev" Users from visit counts.