Make 1st move safe, even for Pattern Games
Previously, we placed mines before the first move for Pattern Games.
This made it possible that the first move could be an Game-losing move.
Instead, we will now disrupt the pattern and turn the seed cell into a
safe Cell as needed.

This felt important since Pattern games are so rare now... sucks to
instantly lose the game without having a chance.

Also, fix the public API exposure on the PlaceMines and
RandomlyPlaceMines service objects.