Refactor `@view` -> `@<action_name>` (view model names)
Instead of using `@view`, which was meant to be a generic term for the
"View Model" for an action... we now use `@<action_name>`, which is a
specific name that actually matches the View Model's name itself. So,
instead of `@view = Home::Show` we now have `@show = Home::Show`, which
is inline with our norm for variable names. This helps because:
1. It is consistent with the norm for variable names
2. It is more illuminating to now the type of the variable than its
   function (... thus our norm)
3. It distances itself from a term that sounds a bit like the concept of
   the "view template" and moves toward an Action-specific model. 👍

I had already started moving in this direction before, this commit just
finishes off the effort.