Show close btn on hover for Flash Notifications
The close button tends to make the text for notifications be
off-centered, and otherwise just takes up valuable space.

Instead, we take a design queue from macOS and just only show the close
button on notification hover. This fixes both issues in a fairly simple
way.

However, since we can't hover on mobile, we just always show the close
button on mobile. To support this, we add a `desktop` variant to our
Tailwind CSS file.

Note: Wwe have to add `:self` to the end of
`transitionend->flash#remove:self` because the transition event bubbles
up from the close button to the parent--the notification container.
Without this, the transition of the child element--the close
button--was removing the notification earlier than wanted.