Fix up styling of flash notifications
Fix centering of flash notifications to not have an awkward amount
of space around them by removing the `left-1/2 -translate-x-1/2` and
adding, instead, `inset-x-0 px-12`. This stretches the container across
the page while leaving an intentional margin buffer on either side. (The
net result is less of a margin buffer and allowing the notification
container to stretch a lot more on wider browser widths--on desktop.)

Also, move the flash notifications container down a bit so it leaves
more of the main nav bar exposed. This feels a bit better for the
overall design on this site.

Finally: Make the sample flash messages use our actual flash messages,
just with dummy text interpolated in.