Fix Fleet Mustering notification line height
The increased font size of the "reload" icon was causing the line height
of the overall notification to be off, giving it a feel of extra padding
around the text as compared to other notifications. I'm still not sure
what the line height should be for base text sizes, but using `5` works
here.