Greatly simplify the Welcome Banner concept/code
Motivation comes from... I don't think people were closing the banner
before. But they don't need it after having played a full game and
signed their name. Also because the close button was making the text
off-centered and ugly on mobile.

So now...
No more "close button" for the Welcome Banner. It'll show for
non-signers and go away for Users that have ever signed before. No
cookie storage, no JS, no complicated contexts, ... wow so much was
able to be removed for just this one "simple" feature.

There is now an old "welcome_banner" cookie that will just sit around
and languish... at least until someone does a "Voluntary Discharge" to
unlink their account. But that's ok.

The only thing I'm not loving now is that the welcome banner just blinks
out of existence after signing for the first time (vs fading away).
But... I think the simplicity of this is worth it. Otherwise, we need
to implement a Turbo "add CSS class" helper, etc. And would then need
to bring back the BannerController JS to remove the DOM element
ontransitionend.