Refactor `turbo_update_target` -> `turbo_target`
We've since moved on to just `turbo_target` for this sort of thing...
So this commit gets us up to date from the old style to the new one.

This idea here is to have a generic enough name that the action being
performed can vary over time without requiring a name change. e.g.
`turbo_update_target` ... changing from an `update` to a `replace` would
seem to require a name change to match. Also, just `turob_target` is
shorter, to boot.