Replace the entire board on Cell Effects
This commit is just a small step towards replacing the entire board on
all Cell Effects (as we once did). There's still surrounding code that
could be optimized for this approach. But this gets us there early for
testing things out.

To support this, we needed to include a portion of what was defined by
Games::Current::Board::Cells::ActionBehaviors. Namely, the `game`,
`board`, and `cell` getter methods. So, we extract this portion out into
just a Games::Current::Board::Cells::Behaviors mix-in. Then, for
parallelism, we define Games::Current::Board::Cells::EffectBehaviors
as the mix-in for the Cell Effect controllers and just include
Games::Current::Board::Cells::Behaviors within that (only... at least
for now).
- The reason we must stop using `Cell.find(params[:cell_id])` is that
  we now need the Cells that we're applying the effects to to be aware
  of their entire Board context since we're rendering back that entire
  Board now. It's a bit more processing, which I was trying to get away
  from for simple effects, but... it's where we are now.

Also, define `generate_game_update_action` in
Games::Current::Board::Cells::DispatchBehaviors instead of separately
in each of:
- Games::Current::Board::Cells::DispatchAction::Dispatch and
- Games::Current::Board::Cells::DispatchEffect::Dispatch
As the behavior is identical and needs to change in lock-step.