Persist Cell highlights to fix highlight "stealing" bug
Long ago, we did persist Cell Highlight details directly in the Cell
object in the database. But when we switched from full-board refreshes
to per-Cell updates (only updating the Cells that changed), we didn't
need this anymore. So, we turned highlights into virtual attributes.

However, now that we're back to full-board refreshes (for a number of
reasons) we have to go back to storing Cell highlight information in
the database. Else, what happens is: Given User A clicks and holds on a
revealed Cell, showing highlights... if User B then does any Cell
action, the board will refresh and wipe out the highlights that User A
is still currently in the middle of. Thus, "stealing" the highlight.

To fix this, we persist highlight info into the cells table. This way,
when the board is refreshed, this info comes with it.

Note: This isn't a simple revert of functionality. We now have
a "highlight origin" effect as well that we have to deal with, where we
didn't used to.
- An interesting aspect of having this Highlight Origin concept is that
  we can actually use it to perform a Toggle Highlight Neighborhood
  effect. Thus, not needing individualized Highlight and Dehighlight
  controllers/actions anymore. So we may come back to this later.