Don't delay for first moves in Auto Solver
Now that we're refreshing the whole board every time, there's no need to
wait longer for the first move in the DEBUG-mode auto-solver. Every
once in a while I'm seeing the need on the last move, so leaving that
part in.

Also, clean up the code a bit... use clearer function names, move
function calls to more sensible locations, etc. Basically cleaning up
what Grok 2 did that wasn't really the best after applying a little
human though to it this time. At the same time, I'm trying to not to
spend too much time on this one since it's already a solved problem and
is only a debug tool.