Always Dehighlight Neighbors onMouseUp
Previously, one could mousedown on a cell to highlight its neighbors,
then drag out of the board and mouseup. This would keep the cell
highlighted "forever".

Now, we change the `mouseup` event handler to be on the document, which
allows us to always compare if the event.target is the original Cell or
not. If not, we call Dehighlight Neighbors, as before.

As an optimization step, since this does mean we'll be hitting this
DesktopController#dispatchMouseup on every click while in the War Room
(home page), we first check if `this.mouseDownTarget` is set. Then we
make sure to always unset it by the end of the dispatchMouseup event.
This last bit is a cleanup step we probably should have before.