Make App.created_at more precise
Making App.created_at more precise doesn't really affect anything. I
just happened to notice it based on some code that won't end up being
checked in that was enumerating a wall of dates between App.created_at
and now. That made me want to reduce that wall of dates for a worst-case
scenario at least.

Aside from that, tracking the actual day, and not just the year, that
the app was created is kind of cool. So, just to encapsulate that
somehow, we also add the created_at date to `App.introspect` (along w/
the App name).