Fix: Allow ObjectInspector helpers in DEBUG mode
Not sure why I originally thought I'd want to not see
ObjectInspector-style #inspect output in DEBUG mode... but I have been
wondering for a long time now why this was the case. I never caught on
to it being a DEBUG mode-only thing since I normally run my server with
DEBUG=1. I was thinking it was something particular to the new debug gem
or something. Yikes!