Load next page content before scrolling to bottom of pg
i.e. prefetch lazy lazy-loaded turbo_frames before they become visible
in the viewport.

The goal here is to smooth out our new infinite scrolling effect on the
Game Index page and User Show page by pre-fetching the lazy-loading
"pager" turbo frame before it comes into view. We can adjust the margin
at which we initiate the next page of results, but the default is 500px.
We use `500px` somewhat arbitrarily... but it seems to behave pretty
well.

We also force the lazy-loading "pager" turbo frame to have a height of
1000px so that the scroll bar is less wildly affected by loading new
content. This number is also somewhat arbitrary. We're just trying to
guesstimate what an average height of newly-loaded content might be.

Reference:
https://radanskoric.com/articles/load-lazy-loaded-frame-before-it-scrolls-in-view