Show User's Time Zone in User Profile page
This helps to understand where different Users are in the world.

Note: There may be some edge case where it's possible for a User to not
have a Time Zone recorded (e.g. if JS is disabled or something). So,
jic, we default to showing the `View.no_value_indicator` if
User#time_zone is nil.