Refactor unneeded instance variables in controllers
We're following the Sandi Metz rule: Only define 1 instance variable
per controller. I hadn't thought of a practical way to do that with
my `before_action :require_<resource>` methods before... but it was
suddenly obvious today. We just define a private `attr_accessor` and
voila!

This change also gives us all the nice effects of using `barewords` in
development. 👍 (Thanks, Avdi!) See:
https://www.alchemists.io/articles/barewords_pattern/

The reason we do this is to protect our public API. And the public API
for view templates/partials is every instance variable defined in the
controller. So if we limit that to 1 instance variable then we maintain
full control with no chance of a view template sneakily side-stepping
that interface. Not that that's going to happen in this app, but it's
all about being used to the style of development.