Show recently ended Games on Home page (for 3 minutes)
It's a bit disorienting/jarring to have just been playing a game and ...
for various reasons, no longer be able to find it. For example:
- If you've navigated away for a second and come back but... since the
  game just ended, now you're on the Home page with New Game buttons
  (only).
- If you've been unlucky enough to hit a race condition where:
  You mousedown on a Cell, but before you can mouseup someone else has
  just ended the game. So when your mouseup event fires, it goes to the
  Cell Action controller and registers as an attempt to revel(Neighbors)
  on the non-Current Game. This then rejects the request and reloads
  the page. And so, suddenly you're on the Home page with the New Game
  buttons thinking: wtf?

To fix this, we make our Game::Current::Find service object consider
current OR recently-ended games.
- We define "recently ended" as "within the last 3 minutes".
Then, on the Home::Show view, we check if the Game is not "on?", in
which case we show the Just-Ended Game screen.

Note: The 3-minute duration for "recently ended" games syncs up with
FleetTracker::DEEP_EXPIRATION_MINUTES on purpose. But, we also don't
tie the two together, on purpose. I'm not sure they should be synced,
but I do like that the timers are at least close. This would mean that
a player that doesn't do anything else after game-end will tend to
"deep expire" / be removed from the Roster around the same time as the
Home page wills tart showing the "New Game" view again. Though there is
some discrepancy here since it may take some time for the recurring job
to fire to clean up the "deeply expired" Roster listings.

Still to come: Reloading the Just-Ended Game screen causes 2 requests
every time still. One for the page and then one for the Results
section. While this is essential for the Game-end transition, it's not
needed for page reloads or fresh Home page visits. So we'll want to
fix this.