Show e.g. "T+:01:34" in Just-ended Game title
It's hard to parse the timestamp ranges in Past Games. While we may not
be able to do something better for Past Games... For Just-ended Games,
we can do better.

If a game actually *Just* ended, we will not show any timestamp /
details.

If a game has "recently" ended, however, we will now show a "T+"-style
timestamp indicating how long ago the Game ended.

This is not dynamically updated (at least not yet). So it requires a
page refresh to update. But it's ever so much more useful!

The price we pay for this is having to break out Just-Ended Game
versions of the Title view model and _title partial. So there's now
some fresh new duplication, but that's OK! It allows us to grow each
of these concerns separately. We can always de-duplicate specific
methods/code later if needed.

Note: Up to this point, the Duration model had actually fallen out of
use at some point in the past. But it's exactly what we needed! So, for
once, I'm glad I didn't properly remove old, unused code! Especially
because, in this case, I wouldn't have remembered that it had ever
existed or what it did or that it was perfect for what we needed. :o