Optimize Game-Over failures on Cell Actions
Instead of failing with a 404 response when a Cell Action is attempted
on a Game-over Game:
1. Go ahead and find the Game, but mark the Cells as "readonly".
2. Raise a Games::Current::Board::Cells::DispatchAction::Error if a
   Cell Action is attempted. This gives us a bit more info in the form
   of the message: `"Game is over: #{game.status.inspect}"`.

The idea is that this will make it easier to distinguish between just
some random attempt to find a Game by ID that doesn't exist, and an
e.g. race condition trying to perform a Cell Action on a Game that has
previously ended.

Also, try to trim down the raw code in
Games::Current::Board::Cells::Behaviors a bit.