Clean up the (Welcome) Banner code + Add "Learn More" link
This was sparked by looking at the wall of text that is the Welcome
Banner on mobile. It was too much. So I wanted to just trim down this
text at least for the home page.

But I've also been wanting to make it easier for people to learn more
about the site. So, in a nod to Starship Troopers, we add a "Would you
like to know more?" link at the end.

Further, by splitting up the site description key in en.yml, we can
use a more piecemeal approach to get just the right message in the right
place, while minimizing duplication.

Also, reduce the overhead of making/showing banners... by removing
Application::Banner. And so on.