On Game end: Replace Game for self + refresh for others
This mixes the 2 styles... for the current User--the one that just ended
the Game--we just instantly replace the Game container via the request
response payload. For all other users, they'll receive the broadcast to
refresh their page.

To make this separation possible, we have to divvy up the
`turbo_stream_actions` collection into "response" actions and
"broadcast" actions. We do this by creating our own, custom
TurboStreamActions object. With this object we can either add turbo
stream actions to both buckets, or: just to the "response" bucket or
just to the "broadcast" bucket. Typically we'd add to both, and the
usage for this doesn't change (`turbo_stream_actions << ...`). Else,
we specify with:
- `turbo_stream_actions.response << ...` or
- `turbo_stream_actions.broadcast << ...`

Then we must update WarROomResponder to expect this new object type and
handle the divvying up for us from there.

Note: While the current user session will receive the broadcast, Turbo
filters it out since it's a `turbo_stream.refresh` action, which does
debounce this action for the current user session. (This is the only
Turbo Stream action that will do this for us.)